Add a projectCache to issue creation, plus specify issueTypes
We default to the first issuetype available to the particular project.
If somehow a project has no available issue types (this might be
impossible, since you need a type to create an issue), then we skip
creating any issues for that project, after logging to stderr.

Additionally, create a helper function to log responses to stderr on
error cases.